pyproject.toml
[project]
name = "agent-programming-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.2",
    "httpx[socks]>=0.28.1",
    "langgraph>=1.0.10",
    "openai>=2.17.0",
    "python-dotenv>=1.2.1",
]
